Move AlbumData to its own class
